refactor of register and login, refactor homepage, change dao to repository

+++ src/main/java/com/gameshop/service/ProductService.java
@@ -8,7 +8,7 @@ import com.gameshop.entity.Product;
 public interface ProductService {
 	public List<Product> getProducts();
 
-	public Product getProductById(int productId);
+	public Product getProductById(Long productId);
 
 	public Set<String> fetchCategoriesFromProducts(List<Product> products);
